% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_manifest.R
\name{annotate_manifest_with_idat}
\alias{annotate_manifest_with_idat}
\title{Add idat information to manifest}
\usage{
annotate_manifest_with_idat(manifest, idat)
}
\arguments{
\item{manifest}{Object with manifest, e.g. from \code{\link[=load_manifest]{load_manifest()}}}

\item{idat}{Object with idat info, e.g. from \code{\link[=read_idat_single]{read_idat_single()}}}
}
\description{
Add idat information to manifest
}
