% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_alleles.R
\name{call_alleles}
\alias{call_alleles}
\title{Call alleles}
\usage{
call_alleles(
  d_lst,
  cutoff,
  nc_int_thres,
  nc_snr_thres,
  nc_sd_thres,
  nc_nbeads_thres
)
}
\arguments{
\item{d_lst}{List of \code{data.frame}'s}

\item{top_n}{Only take \code{top_n} rows, if \code{0} (default), take all rows}
}
\description{
Call alleles
}
