% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_alleles.R
\name{call_alleles_ab_compare}
\alias{call_alleles_ab_compare}
\title{Compare call alleles}
\usage{
call_alleles_ab_compare(
  d_lst,
  cutoff,
  nc_int_thres = NULL,
  nc_snr_thres = NULL,
  nc_sd_thres = NULL,
  nc_nbeads_thres = NULL
)
}
\arguments{
\item{d_lst}{List of \code{data.frame}'s}

\item{top_n}{Only take \code{top_n} rows, if \code{0} (default), take all rows}
}
\description{
Compare call alleles
}
