% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_probe_to_alleles.R
\name{channel_probe_to_alleles}
\alias{channel_probe_to_alleles}
\title{Call alleles}
\usage{
channel_probe_to_alleles(annotated_manifest, top_n = 0)
}
\arguments{
\item{top_n}{Only take \code{top_n} rows, if \code{0} (default), take all rows}

\item{d}{Manifest annotated with idat e.g. from \code{\link[=annotate_manifest_with_idat]{annotate_manifest_with_idat()}}}
}
\description{
Call alleles
}
