% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idat_ex}
\alias{idat_ex}
\title{Example of idat}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 200 rows and 33 columns.
}
\usage{
idat_ex
}
\description{
This exemplify data loaded with \code{\link[=read_idat_single]{read_idat_single()}}.
This is just simulated data.
}
\details{
See vignettes for example of usage.
}
\keyword{datasets}
