% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idat.R
\name{read_idat_single}
\alias{read_idat_single}
\title{Read idat file from red and green channel from sample sheet}
\usage{
read_idat_single(datapath, path_red, path_grn)
}
\arguments{
\item{datapath}{Folder with \code{idat} files}

\item{path_red}{Filename for the \code{idat} file with the red channel}

\item{path_grn}{Filename for the \code{idat} file with the green channel}
}
\value{
\code{data.table} with columns \code{Name}
}
\description{
Read idat file from red and green channel from sample sheet
}
