% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_call.R
\name{call_alleles}
\alias{call_alleles}
\title{Call alleles}
\usage{
call_alleles(annotated_manifest, top_n)
}
\arguments{
\item{top_n}{Only take \code{top_n} rows, if \code{0} (default), take all rows}

\item{d}{Manifest annotated with idat e.g. from \code{\link[=annotate_manifest_with_idat]{annotate_manifest_with_idat()}}}
}
\description{
Call alleles
}
